requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
tkinter-qpython==8.6.0.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
